Try to handle access token
This was my first attempt at maintaining the access token. I couldn't
get this to work because I was getting a 401 error because of invalid
access token. Added a link to notes.txt on how to do it with PKCE flow
that seems to suit my use case. It stores the token in localStorage.
There is also a github link on the link I added