[Tut] Module 2
I needed to scrap everything. The native spotify tutorial is so out of
date I couldn't even run it. There are many issues open in their repo
for it from as far back as 2 years ago. I'm gonna follow the tutorial
linked in notes.txt from now on. It might be out of date, but she
explains everything so at least I wouldnt be in the dark. The commit
messages will be like this from now on till I finish the tutorial so I
can keep my place