[Tut3] Add nodemon
This adds nodemon as a dev dependancy. It allows you to keep the same
instance of index.js running even when changes are made to index.js and
saved. If you change something and save, it'll automatically restart
itself, and refreshing the browser will reflect the change. Like how it
does it in react