[Tut3] Add the Spotify token auth functionality
This code correctly redirects to spotify auth in the /login route. After
login is successful it will redirect back to the /callback route where
it displays the token response json. This should now work regardless of
incogito mode but I'm not sure about that yet. Need to add better
comments and such but I gotta run for now