[Tut3] Add a refresh_token route handler
Route handler for refresh token exists and will work I think. I followed
the tutorial on this so don't fully understand, but when the get for the
profile is replaced with the refresh callback, it works. Idk if this
route automatically gets hit when refresh is needed, or if I have to
handle that in code. Module 3 of tutorial is now done