[Tut4] Improve dev workflow
Add more + better scripts to the root package.json to use concurrently
and postinstall. This means that both the react and express apps will
npm start and npm install together.

Also add a proxy to the client package.json so that any unknown request
made by the client point to the server. This specifically helps for the
refresh token example in useEffect of App.js. The href for logging in is
fine because thats a full redirect, but if the client tried to make a
fetch to the server without the proxy, it would break. You also can't
just refrence the other port because you will get a CORS issue