Add categorization based off tempo
Made the audioFeatures just a part of the track information. This makes
it easier to keep everything together. Made the categorization criteria
for bpm just greater than or less than the desired bpm for now (a
utility function to do the math will be needed later), and added it into
the track data as well. Clicking the button on the result page will show
the structure of this data